# openapi.yaml
openapi: 3.1.0
info:
  title: Plataforma RH Leoni Cable
  description: >
    Plataforma on-premise de Recursos Humanos para Leoni Cable (Mexico).
    Gestiona solicitudes de empleados (vacaciones, permisos), incidencias,
    actas administrativas, comedor/acceso biometrico, organigrama e
    integracion con nomina TRESS.
  version: 1.0.0
  contact:
    name: Leoni Cable - Equipo de Desarrollo
servers:
  - url: http://localhost:8000
    description: Desarrollo local (Docker)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Clave de API para terminales de comedor
  schemas:
...
    AreaResponse:
      type: object
      required: [area_id, descripcion, estatus_id]
      properties:
        area_id:
          type: integer
        descripcion:
          type: string
        estatus_id:
          type: integer
...
    PuestoResponse:
      type: object
      required: [puesto_id, descripcion, estatus_id]
      properties:
        puesto_id:
          type: integer
        area_id:
          type: [integer, "null"]
        descripcion:
          type: string
        estatus_id:
          type: integer
...
    EstadoEmpleadoResponse:
      type: object
      required: [estado_id, descripcion, estatus_id]
      properties:
        estado_id:
          type: integer
        descripcion:
          type: string
        estatus_id:
          type: integer
...
    RolBrief:
      type: object
      required: [id, nombre]
      properties:
        id:
          type: integer
        nombre:
          type: string
...
    UsuarioResponse:
      type: object
      required: [id, empleado_id, no_empleado, nombre, rol_id, created_at]
      properties:
        id:
          type: integer
        empleado_id:
          type: integer
        no_empleado:
          type: string
        nombre:
          type: string
        email:
          type: [string, "null"]
        rol_id:
          type: integer
        rol:
          $ref: "#/components/schemas/RolBrief"
        estado:
          $ref: "#/components/schemas/EstadoEmpleadoResponse"
        area:
          $ref: "#/components/schemas/AreaResponse"
        subarea:
          $ref: "#/components/schemas/SubareaResponse"
        puesto:
          $ref: "#/components/schemas/PuestoResponse"
        categoria:
          $ref: "#/components/schemas/CategoriaResponse"
        clasificacion:
          $ref: "#/components/schemas/ClasificacionEmpleadoResponse"
        lider_id:
          type: [integer, "null"]
          description: empleado_id del líder inmediato
        centrocosto_id:
          type: [integer, "null"]
        foto:
          type: [string, "null"]
        registro:
          type: [string, "null"]
          format: date
        created_at:
          type: string
          format: date-time
    UsuarioListItem:
      allOf:
        - $ref: "#/components/schemas/UsuarioResponse"
        - type: object
          properties:
            lider_nombre:
              type: [string, "null"]
...
    UsuarioPageResponse:
      type: object
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UsuarioListItem"
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
    EmpleadoDistribucionItem:
      type: object
      required: [label, total]
      properties:
        label:
          type: string
        total:
          type: integer
    EmpleadosPorClasificacionAreaSerie:
      type: object
      required: [tipo, clasificacion_descripcion, por_area]
      properties:
        tipo:
          type: string
          enum: [administrativo, directo, indirecto]
        clasificacion_id:
          type: integer
          nullable: true
        clasificacion_descripcion:
          type: string
        por_area:
          type: array
          items:
            $ref: "#/components/schemas/EmpleadoDistribucionItem"
    UsuarioResumenResponse:
      type: object
      required:
        - total_plantilla
        - activos
        - inactivos
        - sin_lider_asignado
        - practicantes
        - porcentaje_operatividad
        - colaboradores_total
        - contratos_por_vencer
        - empleados_por_clasificacion_y_area
      properties:
        total_plantilla:
          type: integer
        activos:
          type: integer
        inactivos:
          type: integer
        sin_lider_asignado:
          type: integer
        sin_email_administrativo:
          type: integer
          description: Administrativos activos sin correo en empleado.email
        practicantes:
          type: integer
        porcentaje_operatividad:
          type: number
          format: float
        colaboradores_total:
          type: integer
        contratos_por_vencer:
          type: integer
        empleados_por_clasificacion_y_area:
          type: array
          items:
            $ref: "#/components/schemas/EmpleadosPorClasificacionAreaSerie"
    CatalogoFiltrosResponse:
      type: object
      required: [areas, puestos]
      properties:
        areas:
          type: array
          items:
            $ref: "#/components/schemas/AreaResponse"
        puestos:
          type: array
          items:
            $ref: "#/components/schemas/PuestoResponse"
    SolicitudBrief:
      type: object
      required: [id, tipo, estado, fecha_inicio, fecha_fin, created_at]
      properties:
        id:
          type: integer
        tipo:
          type: string
        estado:
          type: string
        fecha_inicio:
          type: string
          format: date
        fecha_fin:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
    IncidenciaBrief:
      type: object
      required: [id, tipo, created_at]
      properties:
        id:
          type: integer
        tipo:
          type: string
        estatus_id:
          type: [integer, "null"]
        created_at:
          type: string
          format: date-time
    ActaBrief:
      type: object
      required: [id, estado, created_at]
      properties:
        id:
          type: integer
        estado:
          type: string
        created_at:
          type: string
          format: date-time
    Vista360TurnoEmpleado:
      type: object
      description: Datos de turnos_empleados; la API solo lo incluye cuando el solicitante tiene rol rh.
      properties:
        comedor:
          type: string
          nullable: true
          description: Nombre del comedor asignado (resuelto desde catálogo comedores).
        turno:
          type: string
          nullable: true
    VacacionesResponse:
      type: object
      required:
        - empleado_id
        - dias_disponibles
      properties:
        empleado_id:
          type: integer
        dias_disponibles:
          type: integer
          minimum: 0
        updated_at:
          type: string
          format: date-time
          nullable: true
    VacacionesUpdate:
      type: object
      required:
        - dias_disponibles
      properties:
        dias_disponibles:
          type: integer
          minimum: 0
          description: Saldo de días de vacaciones disponibles
    UsuarioVista360Response:
      type: object
      required:
        - usuario
        - solicitudes_recientes
        - incidencias_activas
        - actas_firmadas
        - saldo_vacaciones
      properties:
        usuario:
          $ref: "#/components/schemas/UsuarioResponse"
        solicitudes_recientes:
          type: array
          items:
            $ref: "#/components/schemas/SolicitudBrief"
        incidencias_activas:
          type: array
          items:
            $ref: "#/components/schemas/IncidenciaBrief"
        actas_firmadas:
          type: array
          items:
            $ref: "#/components/schemas/ActaBrief"
        saldo_vacaciones:
          type: integer
        turno_empleado:
          allOf:
            - $ref: "#/components/schemas/Vista360TurnoEmpleado"
          nullable: true
          description: Presente solo para rol rh; ausente o null para otros roles.
    MetricasUsuarioResponse:
      type: object
      required:
        - solicitudes_por_estado
        - incidencias_por_tipo
        - dias_antiguedad
        - total_actas
      properties:
        solicitudes_por_estado:
          type: object
          additionalProperties:
            type: integer
        incidencias_por_tipo:
          type: object
          additionalProperties:
            type: integer
        dias_antiguedad:
          type: integer
        total_actas:
          type: integer
    ActasPageResponse:
      type: object
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ActaResponse"
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer

    # ── Solicitudes ───────────────────────────────────
...
    ActaResponse:
      type: object
      required: [id, empleado_id, estado, generado_por, created_at]
      properties:
        id:
          type: integer
        empleado_id:
          type: integer
        empleado_nombre:
          type: [string, "null"]
        numero_empleado:
          type: [string, "null"]
        puesto:
          type: [string, "null"]
        area_departamento:
          type: [string, "null"]
        supervisor_directo:
          type: [string, "null"]
        tipo_falta:
          type: [string, "null"]
        fundamento_legal:
          type: [string, "null"]
          enum:
            - Ley Federal del Trabajo
            - Reglamento Interior de Trabajo
            - null
        articulo_inciso:
          type: [string, "null"]
        fecha_evento:
          type: [string, "null"]
          format: date
        lugar_incidente:
          type: [string, "null"]
        descripcion_hechos:
          type: [string, "null"]
        personas_involucradas:
          type: [string, "null"]
        testigos:
          type: [string, "null"]
        responsable_rh:
          type: [string, "null"]
        evidencia:
          type: [string, "null"]
        incidencia_id:
          type: [integer, "null"]
        contenido_ia:
          type: [string, "null"]
        ia_recomendacion:
          type: [string, "null"]
        contenido_final:
          type: [string, "null"]
        estado:
          type: string
          enum: [draft, pending_sign, signed, archived, cancelled]
        generado_por:
          type: integer
        created_at:
          type: string
          format: date-time
        aprobaciones:
          type: array
          items:
            $ref: "#/components/schemas/ActaAprobacionResponse"
          default: []
        firmantes_pendientes:
          type: array
          items:
            type: string
          default: []
...
    GradoPuestoCreate:
      type: object
      required: [nombre, orden]
      properties:
        nombre:
          type: string
          minLength: 2
          maxLength: 100
        orden:
          type: integer
          minimum: 1
          maximum: 99
    GradoPuestoUpdate:
      type: object
      required: [nombre, orden]
      properties:
        nombre:
          type: string
          minLength: 2
          maxLength: 100
        orden:
          type: integer
          minimum: 1
          maximum: 99
    GradoPuestoResponse:
      type: object
      required: [id, nombre, orden, activo, created_at, updated_at]
      properties:
        id:
          type: integer
        nombre:
          type: string
        orden:
          type: integer
        activo:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    GradoPuestoListResponse:
      type: object
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/GradoPuestoResponse"
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
...
    EvaluacionResponse:
      type: object
      required: [id, empleado_id, competencia_id, nivel_actual, fecha_evaluacion, created_at, updated_at]
      properties:
        id:
          type: integer
        empleado_id:
          type: integer
        empleado_nombre:
          type: string
          nullable: true
        competencia_id:
          type: integer
        competencia_nombre:
          type: string
          nullable: true
        nivel_actual:
          type: integer
          minimum: 0
          maximum: 4
        evaluador_id:
          type: integer
          nullable: true
        evaluador_nombre:
          type: string
          nullable: true
        observaciones:
          type: string
          nullable: true
        fecha_evaluacion:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /api/v1/empleados/resumen:
    get:
      tags: [Empleados]
      summary: Resumen de plantilla
      operationId: empleados_resumen
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Metricas de plantilla
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioResumenResponse"
  /api/v1/empleados/catalogo-filtros:
    get:
      tags: [Empleados]
      summary: Catalogos para filtros del directorio
      operationId: empleados_catalogo_filtros
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Areas y puestos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogoFiltrosResponse"
  /api/v1/empleados:
    get:
      tags: [Empleados]
      summary: Directorio de empleados (paginado)
      operationId: empleados_list
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: q
          in: query
          description: Busqueda por nombre o numero de empleado
          schema:
            type: string
        - name: area_id
          in: query
          schema:
            type: integer
        - name: puesto_id
          in: query
          description: Permite multiples valores
          schema:
            type: array
            items:
              type: integer
        - name: activo
          in: query
          description: "Solo RH: true=activos, false=inactivos, omitir=todos"
          schema:
            type: boolean
        - name: estatus
          in: query
          description: "Supervisor/Gerente: activo, inactivo, permiso"
          schema:
            type: string
            enum: [activo, inactivo, permiso]
        - name: solo_contratos_por_vencer
          in: query
          description: Contratos que vencen en 30 dias
          schema:
            type: boolean
        - name: solo_sin_lider
          in: query
          description: "Solo RH: empleados activos sin lider asignado (mismo criterio que KPI)"
          schema:
            type: boolean
        - name: solo_sin_email
          in: query
          description: "Solo RH: administrativos activos sin email registrado (mismo criterio que KPI)"
          schema:
            type: boolean
      responses:
        "200":
          description: Pagina de empleados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioPageResponse"
  /api/v1/empleados/{empleado_id}/vacaciones:
    get:
      tags: [Empleados]
      summary: Saldo de vacaciones del empleado
      operationId: empleados_vacaciones_get
      security:
        - BearerAuth: []
      parameters:
        - name: empleado_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Saldo de días disponibles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VacacionesResponse"
        "403":
          description: Sin permiso para consultar este empleado
        "404":
          description: Empleado no encontrado
    put:
      tags: [Empleados]
      summary: Actualizar saldo de vacaciones (solo RH)
      operationId: empleados_vacaciones_put
      security:
        - BearerAuth: []
      parameters:
        - name: empleado_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VacacionesUpdate"
      responses:
        "200":
          description: Saldo actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VacacionesResponse"
        "403":
          description: Solo RH puede actualizar el saldo
        "404":
          description: Empleado no encontrado
        "422":
          description: Validación de días disponibles
  /api/v1/empleados/{empleado_id}/foto:
    get:
      tags: [Empleados]
      summary: Fotografía del empleado (directorio RH/Images)
      operationId: empleados_foto
      security:
        - BearerAuth: []
      parameters:
        - name: empleado_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Imagen del empleado
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
        "404":
          description: Empleado o fotografía no encontrados
  /api/v1/empleados/{empleado_id}/vista360:
    get:
      tags: [Empleados]
      summary: Vista 360 del empleado
      operationId: empleados_vista360
      security:
        - BearerAuth: []
      parameters:
        - name: empleado_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Perfil completo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioVista360Response"
  /api/v1/empleados/{empleado_id}/metricas:
    get:
      tags: [Empleados]
      summary: Metricas individuales del empleado
      operationId: empleados_metricas
      security:
        - BearerAuth: []
      parameters:
        - name: empleado_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: KPIs del empleado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricasUsuarioResponse"
  /api/v1/empleados/{empleado_id}/actas:
    get:
      tags: [Empleados]
      summary: Actas administrativas del empleado (paginado)
      operationId: empleados_actas_page
      security:
        - BearerAuth: []
      parameters:
        - name: empleado_id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 5
      responses:
        "200":
          description: Actas del empleado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActasPageResponse"

  # ── Solicitudes ─────────────────────────────────────
...
  /api/v1/grados-puesto:
    get:
      tags: [Talento - Grados Puesto]
      summary: Listar grados de puesto (paginado)
      operationId: grados_puesto_list
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
        - name: busqueda
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Pagina de grados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GradoPuestoListResponse"
    post:
      tags: [Talento - Grados Puesto]
      summary: Crear grado de puesto
      operationId: grados_puesto_create
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradoPuestoCreate"
      responses:
        "201":
          description: Grado creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GradoPuestoResponse"
  /api/v1/grados-puesto/{id}:
    get:
      tags: [Talento - Grados Puesto]
      summary: Obtener grado de puesto
      operationId: grados_puesto_get
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Detalle del grado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GradoPuestoResponse"
    patch:
      tags: [Talento - Grados Puesto]
      summary: Actualizar grado de puesto
      operationId: grados_puesto_update
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradoPuestoUpdate"
      responses:
        "200":
          description: Grado actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GradoPuestoResponse"
    delete:
      tags: [Talento - Grados Puesto]
      summary: Eliminar grado de puesto (soft delete)
      operationId: grados_puesto_delete
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Grado eliminado

  # ── Talento — Metodos Calificacion Competencia ───────
...
  /api/v1/competencias/multihabilidades/puestos:
    get:
      tags: [Talento - Competencias]
      summary: Listar puestos para matriz de multihabilidades
      operationId: competencias_multihabilidades_puestos
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Puestos con competencias configuradas
...
  /api/v1/evaluaciones/empleado/{empleado_id}:
    get:
      tags: [Talento - Evaluaciones]
      summary: Evaluaciones de un empleado
      description: Empleado ve las propias; supervisor ve su area; RH ve todas.
      operationId: evaluaciones_by_empleado
      security:
        - BearerAuth: []
      parameters:
        - name: empleado_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Lista de evaluaciones del empleado
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EvaluacionResponse"
        "403":
          description: Sin permiso para ver evaluaciones de este empleado
...